requirement.txt
alembic==1.7.6
aniso8601==9.0.1
anyio==3.4.0
asgiref==3.4.1
async==0.6.2
async-exit-stack==1.0.1
async-generator==1.10
autopep8==1.6.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.71.0
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.0
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
orjson==3.6.5
passlib==1.7.4
promise==2.3
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
Rx==3.2.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
